Refactor shipment weight calculations
